reduce warning.

+++ dpu/gc.h
@@ -35,20 +35,20 @@ void free_node(mL3ptr nn, int size) {
 }
 
 
-pptr alloc_node(int height, int n) { // return (num_node_get, gcnodeaddr)
+pptr alloc_node(int height, uint32_t n) { // return (num_node_get, gcnodeaddr)
     #ifndef GC_ACTIVE
     return PPTR(0, INVALID_DPU_ADDR);
     #else
     IN_DPU_ASSERT(height < MAX_L3_HEIGHT, "alloc node invalid size");
     IN_DPU_ASSERT(n > 0, "alloc node invalid n");
     pptr addr = free_list_l3node[height].size_addr;
-    if (addr.id >= n) { // get more nodes than requied
+    if (addr.id >= n) { // have more nodes than requied
         // gcnode curnode = free_list_l3node[height];
         mgcptr nxt = (mgcptr)free_list_l3node[height].size_addr.addr;
         pptr ret = PPTR(n, nxt);
-        for (int i = 0; i < n; i ++) {
+        for (uint32_t i = 0; i < n; i ++) {
             gcnode nxtnode = *nxt;
-            IN_DPU_ASSERT(nxt != INVALID_DPU_ADDR, "alloc invalid nxt");
+            IN_DPU_ASSERT((uint32_t)nxt != INVALID_DPU_ADDR, "alloc invalid nxt");
             nxt = (mgcptr)nxtnode.size_addr.addr;
         }
         free_list_l3node[height].size_addr = PPTR(addr.id - n, nxt);
